move a bunch more out of server
