+++ packages/server/index.html
@@ -201,7 +201,6 @@
   {{{fullStoryScript}}}
   {{{notificationScript}}}
   {{{showModalScript}}}
-  {{{intercomScript}}}
   {{{segmentScript}}}
   {{{iterateScript}}}
   {{{userScript}}}
