+++ packages/server/lib/embeds/appcues.js
@@ -1,0 +1,4 @@
+module.exports = ({ isProduction }) =>
+  isProduction
+    ? '<script id="appcues-js" src="//fast.appcues.com/49463.js" async></script>'
+    : '';
