+++ packages/server/lib/embeds/bugsnag.js
@@ -1,0 +1,5 @@
+const { getBugsnagScript } = require('../bugsnag');
+
+module.exports = ({ isProduction, userId }) => {
+  return isProduction ? getBugsnagScript(userId) : '';
+};
